Redesign!
This is the third major version of the official thoughtram website,
which has been created in collaboration with some of our closest
friends.

Big thanks go out to

@timche - Tim Cheung was involved in the initial design and inspiration process
for the new website. He understood from the very beginning what we where striving
for and brought a lot of "freshness" to the new design.

@tim_hartmann_ - Tim Hartmann was a key contributor by helping out for the last
two miles so we could reach the goal in time. Tim has excellent CSS skills and made sure
our new website is in good shape.

Without these people, this major update wouldn't be as great as it is.

Team #thoughtramAndFriends